validateBody, validateUser, contact model refactor
