+++ routes/api/authorize.js
@@ -1,7 +1,10 @@
 const express = require('express');
+const validateBody = require('../../middleWares/validateBody');
+const { validateUser } = require('../../middleWares/authorizeMiddlewares');
+
 const router = express.Router();
 
-router.post("/");
+router.post("/", validateBody, validateUser);
 // router.post('/users/login');
 
 module.exports = router;
